Fix broken memory view link in timeline template
This commit fixes a BuildError that occurred when attempting to view memory details from the timeline page.
The issue was caused by an incorrect URL endpoint in the timeline.html template.

Changes made:
- Updated the URL endpoint from 'memories.view' (which doesn't exist) to 'memories.view_memory'
- Fixed the parameter name from 'id' to 'memory_id' to match the route definition

This resolves the routing exception: 'werkzeug.routing.exceptions.BuildError: Could not build url for endpoint memories.view'